ai-engine/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
pydantic
python-multipart
Pillow
scikit-learn
numpy
aiofiles
a2wsgi
